+++ src/api/types/modules.ts
@@ -19,16 +19,19 @@ export interface Feature {
 export interface FeatureQuery {
     feature_id: string;
     source: string;
+    species: string;
 }
 
 export interface FeaturesQuery {
     feature_id: string[];
     source: string;
+    species: string;
 }
 
 export interface FeatureSearchQuery extends Query {
     query: string | string[];
     source?: string;
+    species?: string;
 }
 
 export interface FeatureAutocompleteQuery extends FeatureSearchQuery {