+++ src/tests/mock.ts
@@ -1,4 +1,4 @@
-import * as angular from 'angular';
+import {UploadEvent, UploadEventType} from '../core/services/api';
 import 'ng-file-upload';
 
 
@@ -10,7 +10,7 @@ import 'ng-file-upload';
  * @param fileUID Unique file identifier
  */
 export interface MockUploadHandler<T> {
-    (data: any, fileUID: string): angular.IHttpPromiseCallbackArg<T>;
+    (data: any, fileUID: string): T;
 }
 
 /**
@@ -18,22 +18,23 @@ export interface MockUploadHandler<T> {
  */
 export class MockApiService {
     // Mock upload handler.
-    private _uploadHandler: MockUploadHandler<any> = () => { return <angular.IHttpResponse<string>> {data: null}; }
+    private _uploadHandler: MockUploadHandler<any> = () => ({ data: null });
 
     /**
      * Performs a mock data upload.
      */
-    public upload<T>(data: any, fileUID: string = ''): angular.angularFileUpload.IUploadPromise<T> {
-        // TODO: Augment the promise to enable upload-specific functions (or make them noops).
-        return <angular.angularFileUpload.IUploadPromise<T>> new Promise<angular.IHttpPromiseCallbackArg<T>>(
-            (resolve, reject) => {
-                try {
-                    resolve(this._uploadHandler(data, fileUID));
-                } catch (error) {
-                    reject(error);
-                }
+    public upload<T>(data: any, fileUID: string = ''): Rx.Observable<UploadEvent<T>> {
+        return Rx.Observable.create<UploadEvent<T>>((observer) => {
+            try {
+                observer.onNext({
+                    result: this._uploadHandler(data, fileUID),
+                    type: UploadEventType.RESULT,
+                });
+                observer.onCompleted();
+            } catch (error) {
+                observer.onError(error);
             }
-        );
+        });
     }
 
     /**