+++ docs/CHANGELOG.rst
@@ -9,6 +9,9 @@ This project adheres to `Semantic Versioning <http://semver.org/>`_.
 Unreleased
 ==========
 
+Changed
+-------
+- **BACKWARD INCOMPATIBLE:** Make species part of the feature primary key
 
 ==================
 1.0.0 - 2017-10-24