+++ docs/CHANGELOG.rst
@@ -5,6 +5,14 @@ Change Log
 All notable changes to this project are documented in this file.
 This project adheres to `Semantic Versioning <http://semver.org/>`_.
 
+==========
+Unreleased
+==========
+
+Fixed
+-----
+- Made component loading spinner consistent across angular-material versions
+
 ==================
 2.0.4 - 2017-11-06
 ==================