+++ docs/CHANGELOG.rst
@@ -6,6 +6,14 @@ All notable changes to this project are documented in this file.
 This project adheres to `Semantic Versioning <http://semver.org/>`_.
 
 ==================
+Unreleased
+==================
+
+Fixed
+-----
+- isPromise utility function
+
+==================
 2.0.5 - 2017-11-08
 ==================